finde

+++ Curso Python/07-Ejercicios/Ejercicio8.py
@@ -4,16 +4,10 @@ Cuanto es el x porciento de x numreros
 
 
 """
-# print()
-
-# opcion = 0
-
-# while True:
-
 print("\n")
-numero = int(input("Introduce Un numero: "))
+numero = float(input("Introduce Un numero: "))
 print(" ")
-porcentaje = int(input(f"Que porcentaje quieres sacar de {numero}?: "))
+porcentaje = float(input(f"Que porcentaje quieres sacar de {numero}?: "))
 print(" ")
 operacion = (numero * (porcentaje/100))
 print(f"El {porcentaje}% de {numero} es {operacion} ")
